% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.MUSCLE}
\alias{.MUSCLE}
\title{idealization with MUSCLE}
\usage{
.MUSCLE(Y, q, b, test)
}
\arguments{
\item{Y}{Raw data}

\item{q}{quantiles, can be simulated by simulQuantile_MUSCLE}

\item{b}{target quantile, b = 0.5 for median}

\item{test}{fast test or not, only used by splitting}
}
\description{
idealization with MUSCLE
}
