% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.simulQuantile_DMUSCLE2}
\alias{.simulQuantile_DMUSCLE2}
\title{simulate MUSCLE quantiles}
\usage{
.simulQuantile_DMUSCLE2(p, n, ACF, L)
}
\arguments{
\item{p}{1-alpha}

\item{n}{sample size}

\item{ACF}{estimated autocorrealtion function}

\item{L}{Cholosky decompositon of covariance matrix}
}
\description{
simulate MUSCLE quantiles
}
