% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.simulQuantile_MUSCLE}
\alias{.simulQuantile_MUSCLE}
\title{simulate MUSCLE quantiles}
\usage{
.simulQuantile_MUSCLE(p, n, b)
}
\arguments{
\item{p}{1-alpha}

\item{n}{sample size}

\item{b}{target quantile, b = 0.5 for median}
}
\description{
simulate MUSCLE quantiles
}
