% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUSCLE-package.R
\name{eval_StepFun_M}
\alias{eval_StepFun_M}
\title{Compute segmentation result (only for MMUSCLE).}
\usage{
eval_StepFun_M(stepF, m)
}
\arguments{
\item{stepF}{segmentation result.}

\item{m}{index of segmentation result.}
}
\description{
Compute segmentation result (only for MMUSCLE).
}
