% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUSCLE-package.R
\name{evalStepFun}
\alias{evalStepFun}
\title{evaluate step function
Compute segmentation result.}
\usage{
evalStepFun(stepF)
}
\arguments{
\item{stepF}{segmentation result.}
}
\description{
evaluate step function
Compute segmentation result.
}
