% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUSCLE-package.R
\name{Local_errors}
\alias{Local_errors}
\title{Compute localization errors of estimated change points.}
\usage{
Local_errors(left, left_hat, n)
}
\arguments{
\item{left}{true change points.}

\item{left_hat}{estimated change points.}

\item{n}{sample size.}
}
\description{
Compute localization errors of estimated change points.
}
