% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUSCLE-package.R
\name{MMUSCLE}
\alias{MMUSCLE}
\title{Segmentation with MMUSCLE.}
\usage{
MMUSCLE(
  Y,
  q_matrix,
  beta_vec = c(0.25, 0.5, 0.75),
  implementation = "PRT"
)
}
\arguments{
\item{Y}{observations.}

\item{q_matrix}{quantile matrix, can be simulated by simulQuantile_MMUSCLE.}

\item{beta_vec}{target quantile vector, beta = 0.25, 0.5 and 0.75 stand for first quartile, median and third quartile respectively.}

\item{implementation}{backend implementation: \code{"PRT"} (the default), \code{"PST"}, or \code{"original"}.}
}
\value{
multiple multiscale quantiles segmentation results.
}
\description{
Segmentation with MMUSCLE.
}
