% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUSCLE-package.R
\name{MUSCLE}
\alias{MUSCLE}
\title{Segmentation with MUSCLE.}
\usage{
MUSCLE(
  Y,
  q,
  beta = 0.5,
  dyadic = TRUE,
  split = FALSE,
  m = 2000,
  details = FALSE,
  deconv = FALSE,
  lag = 192,
  implementation = "PRT"
)
}
\arguments{
\item{Y}{observations.}

\item{q}{quantiles, can be simulated by simulQuantile_MUSCLE.}

\item{beta}{target quantile, beta = 0.5 for median.}

\item{dyadic}{boolian variable, indicates test statistics obtained from dyadic subintervals or on full subintervals.}

\item{split}{boolian variable, indicates whether the data will be split by subsets with size m.}

\item{m}{splitting size.}

\item{details}{boolian variable, indicates whether computation details will be displayed.}

\item{deconv}{boolian variable, indicates whether the deconvolution results will be computed.}

\item{lag}{shift lag (the size of dependency), only for deconvolution (deconv = TRUE).}

\item{implementation}{backend implementation: \code{"PRT"} (the default), \code{"PST"}, or \code{"original"}.}
}
\value{
multiscale quantile segmentation results.
}
\description{
Segmentation with MUSCLE.
}
