% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_1000.R
\docType{data}
\name{q_1000}
\alias{q_1000}
\title{Quantile n = 1000}
\format{
A vector with size 1000:
\describe{
  \item{Simulated quantiles n = 1000}{}
}
}
\usage{
q_1000
}
\description{
Simulated quantiles n = 1000
}
\keyword{datasets}
