% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_10000.R
\docType{data}
\name{q_10000}
\alias{q_10000}
\title{Quantile n = 10000}
\format{
A vector with size 10000:
\describe{
  \item{Simulated quantiles n = 10000}{}
}
}
\usage{
q_10000
}
\description{
Simulated quantiles n = 10000
}
\keyword{datasets}
