% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_15000.R
\docType{data}
\name{q_15000}
\alias{q_15000}
\title{Quantile n = 15000}
\format{
A vector with size 15000:
\describe{
  \item{Simulated quantiles n = 15000}{}
}
}
\usage{
q_15000
}
\description{
Simulated quantiles n = 15000
}
\keyword{datasets}
