% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_2000.R
\docType{data}
\name{q_2000}
\alias{q_2000}
\title{Quantile n = 2000}
\format{
A vector with size 2000:
\describe{
  \item{Simulated quantiles n = 2000}{}
}
}
\usage{
q_2000
}
\description{
Simulated quantiles n = 2000
}
\keyword{datasets}
