% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_20000.R
\docType{data}
\name{q_20000}
\alias{q_20000}
\title{Quantile n = 20000}
\format{
A vector with size 20000:
\describe{
  \item{Simulated quantiles n = 20000}{}
}
}
\usage{
q_20000
}
\description{
Simulated quantiles n = 20000
}
\keyword{datasets}
