% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_25000.R
\docType{data}
\name{q_25000}
\alias{q_25000}
\title{Quantile n = 25000}
\format{
A vector with size 25000:
\describe{
  \item{Simulated quantiles n = 25000}{}
}
}
\usage{
q_25000
}
\description{
Simulated quantiles n = 25000
}
\keyword{datasets}
