% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_30000.R
\docType{data}
\name{q_30000}
\alias{q_30000}
\title{Quantile n = 30000}
\format{
A vector with size 30000:
\describe{
  \item{Simulated quantiles n = 30000}{}
}
}
\usage{
q_30000
}
\description{
Simulated quantiles n = 30000
}
\keyword{datasets}
