% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_30000_0-8_0-5.R
\docType{data}
\name{q_30000_0_8_0_5}
\alias{q_30000_0_8_0_5}
\title{Simulated quantiles}
\format{
A vector with length n = 30000:
\describe{
  \item{}{Simulated quantiles with n = 30000, alpha = 0.8, beta = 0.5}
}
}
\usage{
q_30000_0_8_0_5
}
\description{
Simulated quantiles with n = 30000, alpha = 0.8, beta = 0.5
}
\keyword{datasets}
