% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_5000.R
\docType{data}
\name{q_5000}
\alias{q_5000}
\title{Quantile n = 5000}
\format{
A vector with size 5000:
\describe{
  \item{Simulated quantiles n = 5000}{}
}
}
\usage{
q_5000
}
\description{
Simulated quantiles n = 5000
}
\keyword{datasets}
