\name{umax}
\docType{data}
\alias{umax}
\title{Unaligned MAX sequences}
\description{
Unaligned DNA sequences of the MAX gene for 31 mammalian species.
}
\usage{
umax
}
\format{An object of class \code{fasta}.}
\source{\url{http://www.ensembl.org/index.html}}
\references{
Wagner, AJ et al. (1992) Expression, regulation, and chromosomal localization of the Max gene. \emph{Proc Natl Acad Sci USA} \bold{89}: 3111-3115.
}
\keyword{datasets}

