% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUSCLE-package.R
\name{V}
\alias{V}
\title{Compute V-measure of change points estimation.}
\usage{
V(left, left_hat, n)
}
\arguments{
\item{left}{true change points.}

\item{left_hat}{estimated change points.}

\item{n}{sample size}
}
\description{
Compute V-measure of change points estimation.
}
