% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUSCLE-package.R
\name{FDR}
\alias{FDR}
\title{FDR
Compute false discovery rate (FDR) of change points estimation.}
\usage{
FDR(left, left_hat, n)
}
\arguments{
\item{left}{true change points.}

\item{left_hat}{estimated change points.}

\item{n}{sample size}
}
\description{
FDR
Compute false discovery rate (FDR) of change points estimation.
}
