% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUSCLE-package.R
\name{logg}
\alias{logg}
\title{logg
Returns log(x) for positive x and zero for x = 0.}
\usage{
logg(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
logg
Returns log(x) for positive x and zero for x = 0.
}
