% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUSCLE-package.R
\name{MMUSCLE}
\alias{MMUSCLE}
\title{MMUSCLE}
\usage{
MMUSCLE(Y, q_matrix, b_vec = c(0.5))
}
\arguments{
\item{Y}{Raw data}

\item{q_matrix}{quantiles, can be simulated by simulQuantile_MUSCLE}

\item{b_vec}{target quantile, b = 0.5 for median}
}
\value{
idealization of MUSCLE
}
\description{
MMUSCLE
}
