% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUSCLE-package.R
\name{MMUSCLE}
\alias{MMUSCLE}
\title{MMUSCLE
Segmentation with MMUSCLE.}
\usage{
MMUSCLE(Y, q_matrix, beta_vec = c(0.25, 0.5, 0.75))
}
\arguments{
\item{Y}{observations.}

\item{q_matrix}{quantile matrix, can be simulated by simulQuantile_MMUSCLE.}

\item{beta_vec}{target quantile vector, beta = 0.25, 0.5 and 0.75 stand for first quartile, median and third quartile respectively.}
}
\value{
multiple multiscale quantiles segmentation results.
}
\description{
MMUSCLE
Segmentation with MMUSCLE.
}
\examples{
# example code
library(muscle)
# generate data
n = 2048
alpha = 0.1
beta = 0.5
df = 3
#changepoints of the block signal
blocks <- rep(c(0, 14.64, -3.66, 7.32, -7.32, 10.98, -4.39, 3.29, 19.03,
7.68, 15.37, 0), times =c(204, 62, 41, 164, 40, 308, 82, 430, 225, 41, 61,390))
blocks.cpt <- c(205, 267, 308, 472, 512, 820, 902,1332, 1557, 1598, 1659)

signal = blocks
signal.cpt = sort(c(blocks.cpt,390,666,1445))
sd = c(8,0.5,4,1)
set.seed(96)
noise = c(rt(390,df)*sd[1],rt(278,df)*sd[2],rt(779,df)*sd[3],rt(601,df)*sd[4])/sqrt(df/(df-2))
Y = signal + noise

## M-MUSCLE
beta_vec = c(0.25,0.5,0.75)
# simulate quantiles
q_mmuscle = simulQuantile_MMUSCLE(n,alpha = alpha, beta_vec = beta_vec)

# segmentation with M-MUSCLE
# this can be slow!
reg_mmuscle = MMUSCLE(Y,q_mmuscle,beta_vec)

plot(1:n,Y, type = "l", lwd = 2, col = "gray",ylim = c(-40,35),xlab = "",
ylab = "", main = "M-MUSCLE")
lines(1:n, signal, type = "s", lwd =2, col ="black")
lines(eval_StepFun_M(reg_mmuscle,2),type = "s",lwd = 2, col = "red")
lines(eval_StepFun_M(reg_mmuscle,1),type = "s",lwd = 2, col = "orange")
lines(eval_StepFun_M(reg_mmuscle,3),type = "s",lwd = 2, col = "orange")
abline(v = c(390,666,1445), lty = 2, col= "green", lwd = 2 )

}
