% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUSCLE-package.R
\name{MUSCLE}
\alias{MUSCLE}
\title{MUSCLE}
\usage{
MUSCLE(
  Y,
  q,
  b = 0.5,
  dyadic = TRUE,
  split = FALSE,
  m = 2000,
  deconv = FALSE,
  lag = 192
)
}
\arguments{
\item{Y}{Raw data}

\item{q}{quantiles, can be simulated by simulQuantile_MUSCLE}

\item{b}{target quantile, b = 0.5 for median}

\item{dyadic}{test on dyadic subintervals or full subintervals}

\item{split}{boolian variable indicates whether the data will be split by subsets with size m}

\item{m}{splitting size}

\item{deconv}{boolian variable, indicates whether deconvolution or not}

\item{lag}{shift lag, only for deconvolution}
}
\value{
idealization of MUSCLE
}
\description{
MUSCLE
}
