% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUSCLE-package.R
\name{OER}
\alias{OER}
\title{Over estimation rate}
\usage{
OER(left, left_hat)
}
\arguments{
\item{left}{left boundaries of true signals}

\item{left_hat}{left boundaries of etimation}
}
\description{
Over estimation rate
}
