% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUSCLE-package.R
\name{OER}
\alias{OER}
\title{Over estimation rate
Compute over estimation rate (OER) of change points estimation.}
\usage{
OER(left, left_hat)
}
\arguments{
\item{left}{true change points.}

\item{left_hat}{estimated change points.}
}
\description{
Over estimation rate
Compute over estimation rate (OER) of change points estimation.
}
