% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUSCLE-package.R
\name{simulQuantile_MMUSCLE}
\alias{simulQuantile_MMUSCLE}
\title{Simulate MMUSCLE quantiles
Simulate MMUSCLE quantiles.}
\usage{
simulQuantile_MMUSCLE(n, alpha = 0.1, beta_vec = c(0.25, 0.5, 0.75))
}
\arguments{
\item{n}{sample size.}

\item{alpha}{type I error of each multiscale test.}

\item{beta_vec}{target quantile vector, target quantile vector, beta = 0.25,
0.5 and 0.75 stand for first quartile, median and third quartile respectively.}
}
\description{
Simulate MMUSCLE quantiles
Simulate MMUSCLE quantiles.
}
