% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUSCLE-package.R
\name{simulQuantile_MUSCLE}
\alias{simulQuantile_MUSCLE}
\title{Simulate MUSCLE quantiles}
\usage{
simulQuantile_MUSCLE(
  n,
  alpha = 0.1,
  b = 0.5,
  lambda = 1,
  deconv = FALSE,
  Kernel = NULL,
  Y = NULL,
  E16 = FALSE,
  sr = 20000,
  m = 192
)
}
\arguments{
\item{n}{sample size}

\item{alpha}{Type I error of individual test}

\item{b}{target quantile, b = 0.5 for median}

\item{lambda}{regularization parameter}

\item{deconv}{logical variable, TRUE for deconvolution (DMUSCLE)}

\item{Kernel}{convolution kernel}

\item{Y}{data}

\item{E16}{logical variable, TRUE for Element-16 measurement}

\item{sr}{sample rate}

\item{m}{lag}
}
\value{
simulated quantiles
}
\description{
Simulate MUSCLE quantiles
}
