% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUSCLE-package.R
\name{simulQuantile_MUSCLE}
\alias{simulQuantile_MUSCLE}
\title{Simulate MUSCLE quantiles}
\usage{
simulQuantile_MUSCLE(
  n,
  alpha = 0.1,
  beta = 0.5,
  exact = FALSE,
  lambda = 1,
  deconv = FALSE,
  Kernel = NULL,
  ACF = NULL,
  Y = NULL,
  E16 = FALSE,
  sr = 20000,
  lag = 192
)
}
\arguments{
\item{n}{sample size.}

\item{alpha}{type I error of each multiscale test.}

\item{beta}{target quantile, beta = 0.5 for median regression.}

\item{exact}{boolean variable, if exact = TRUE, quantiles will be simulated;
otherwise, quantiles will be computed via appprximation}

\item{lambda}{regularization parameter, only required for deconvolution.}

\item{deconv}{boolean variable, deconv = TRUE for deconvolution (DMUSCLE).}

\item{Kernel}{convolution kernel, only required for deconvolution.}

\item{ACF}{autocorrelation function, only required for deconvolution.}

\item{Y}{observations, only required for deconvolution.}

\item{E16}{A logical variable, TRUE for Element-16 measurement, only required for deconvolution.}

\item{sr}{sampling rate, default value is 20k, only required for deconvolution.}

\item{lag}{shift lag (the size of dependency), only for deconvolution (deconv = TRUE).}
}
\value{
Simulated quantiles.
}
\description{
Simulate MUSCLE quantiles
}
