% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUSCLE-package.R
\name{split}
\alias{split}
\title{split
Split n observations into pieces with size m}
\usage{
split(n, m)
}
\arguments{
\item{n}{sample size}

\item{m}{splitting size}
}
\description{
split
Split n observations into pieces with size m
}
