% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUSCLE-package.R
\name{teethfun}
\alias{teethfun}
\title{teeth function}
\usage{
teethfun(n, K, h = 3)
}
\arguments{
\item{n}{sample size.}

\item{K}{number of change points.}

\item{h}{higher level.}
}
\description{
teeth function
}
