% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUSCLE-package.R
\name{V}
\alias{V}
\title{V-measure}
\usage{
V(left, left_hat, n)
}
\arguments{
\item{left}{left boundaries of true signals}

\item{left_hat}{left boundaries of etimation}

\item{n}{sample size}
}
\description{
V-measure
}
